requirements.txt
djangorestframework
markdown
django-filter
redis
celery
django-celery-results
django-celery-beat